+++ packages/terra-functional-testing/src/commands/axe/index.js
@@ -1,7 +1,5 @@
-const injectAxe = require('./inject');
 const runAxe = require('./run');
 
 module.exports = {
   runAxe,
-  injectAxe,
 };