+++ packages/terra-functional-testing/src/commands/utils/dispatchCustomEvent.js
@@ -3,11 +3,11 @@
  * @param {string} options.name - name of event
  * @param {Object} options.metaData - metadata pertaining to event
  */
-const dispatchCustomEvent = (options) => {
+const dispatchCustomEvent = async (options) => {
   const { name, metaData } = options;
   try {
     // eslint-disable-next-line prefer-arrow-callback
-    global.browser.execute(function dispatchEvent(eventName, eventMetaData) {
+    await global.browser.execute(function dispatchEvent(eventName, eventMetaData) {
       /* If IE support is removed, convert below to use event constructors. */
       const event = document.createEvent('Event');
       event.initEvent(eventName, true, true);