+++ packages/terra-functional-testing/src/commands/utils/setApplicationLocale.js
@@ -7,7 +7,7 @@ const logger = new Logger({ prefix: '[terra-functional-testing:setApplicationLoc
  * Updates terra-application's locale via custom event
  * @param {string} locale - name of locale to update to
  */
-const setApplicationLocale = (locale) => {
+const setApplicationLocale = async (locale) => {
   if (typeof locale === 'undefined' || !locale) {
     logger.error(`Unsupported locale supplied to setApplicationLocale. "${locale}" is not a valid locale.`);
     return;
@@ -18,7 +18,7 @@ const setApplicationLocale = (locale) => {
     metaData: { locale },
   };
 
-  dispatchCustomEvent(event);
+  await dispatchCustomEvent(event);
 };
 
 module.exports = setApplicationLocale;