+++ packages/terra-functional-testing/src/commands/validates/accessibility.js
@@ -2,8 +2,8 @@
  * Analyzes the current documentation for accessibility violations.
  * @param {Object} options - Optional axe configuration overrides.
  */
-function accessibility(options = {}) {
-  expect().toBeAccessible(options);
+async function accessibility(options = {}) {
+  await expect().toBeAccessible(options);
 }
 
 module.exports = accessibility;