+++ packages/terra-functional-testing/src/commands/validates/element.js
@@ -12,15 +12,15 @@ const screenshot = require('./screenshot');
  * @param {number} [options.mismatchTolerance] - the mismatch tolerance for the screenshot comparison.
  * @param {string} [options.selector] - the element selector to use for the screenshot comparison.
  */
-const element = (testName, options = {}) => {
+const element = async (testName, options = {}) => {
   if (!testName || typeof testName !== 'string' || testName.length === 0) {
     throw new Error('[terra-functional-testing:element] Terra.validate.element requires a unique test name as the first argument.');
   }
 
   const { rules, mismatchTolerance, selector } = options;
 
-  accessibility({ rules });
-  screenshot(testName, { mismatchTolerance, selector });
+  await accessibility({ rules });
+  await screenshot(testName, { mismatchTolerance, selector });
 };
 
 module.exports = element;