+++ packages/terra-functional-testing/tests/jest/commands/utils/hideInputCaret.test.js
@@ -3,7 +3,7 @@ jest.mock('@cerner/terra-cli/lib/utils/Logger');
 const { hideInputCaret } = require('../../../../src/commands/utils');
 
 describe('hideInputCaret', () => {
-  it('should hide caret on selector', () => {
+  it('should hide caret on selector', async () => {
     const mockIsExisting = jest.fn().mockImplementation(() => true);
     const element = {
       isExisting: mockIsExisting,
@@ -16,12 +16,12 @@ describe('hideInputCaret', () => {
       execute: mockExecute,
     };
 
-    hideInputCaret('body');
+    await hideInputCaret('body');
 
     expect(mockExecute).toHaveBeenCalledWith('document.querySelector("body").style.caretColor = "transparent";');
   });
 
-  it('should log message when selector is not found', () => {
+  it('should log message when selector is not found', async () => {
     jest.spyOn(console, 'error').mockImplementationOnce(() => { });
     const mockIsExisting = jest.fn().mockImplementation(() => false);
     const element = {
@@ -34,7 +34,7 @@ describe('hideInputCaret', () => {
       execute: mockExecute,
     };
 
-    hideInputCaret('test-selector');
+    await hideInputCaret('test-selector');
 
     expect(mockExecute).not.toHaveBeenCalled();
   });