+++ packages/terra-functional-testing/tests/jest/services/wdio-visual-regression-service/modules/getTerraFormFactor.test.js
@@ -1,7 +1,7 @@
 const getTerraFormFactor = require('../../../../../src/services/wdio-visual-regression-service/modules/getTerraFormFactor');
 
 describe('getTerraFormFactor', () => {
-  it('should return the small viewport', () => {
+  it('should return the small viewport', async () => {
     const size = {
       screenWidth: 622,
       screenHeight: 768,
@@ -11,12 +11,12 @@ describe('getTerraFormFactor', () => {
       execute: () => size,
     };
 
-    const result = getTerraFormFactor();
+    const result = await getTerraFormFactor();
 
     expect(result).toEqual('small');
   });
 
-  it('should return huge as the default viewport size', () => {
+  it('should return huge as the default viewport size', async () => {
     const size = {
       screenWidth: 3000,
       screenHeight: 768,
@@ -26,7 +26,7 @@ describe('getTerraFormFactor', () => {
       execute: () => size,
     };
 
-    const result = getTerraFormFactor();
+    const result = await getTerraFormFactor();
 
     expect(result).toEqual('enormous');
   });