+++ packages/terra-functional-testing/tests/wdio/terra-describe-tests-spec.js
@@ -1,6 +1,6 @@
 Terra.describeTests('Terra.describeTests', { formFactors: ['small', 'huge'], locales: ['en', 'fr'], themes: ['terra-default-theme', 'orion-fusion-theme'] }, () => {
-  it('should filter test by formFactors, locales, and themes', () => {
-    browser.url('/accessible.html');
-    Terra.validates.accessibility();
+  it('should filter test by formFactors, locales, and themes', async () => {
+    await browser.url('/accessible.html');
+    await Terra.validates.accessibility();
   });
 });